Created DataProcessing class plus two new test classes
Issue#252 set up two new testing classes but have not yet implemented any tests.
Added new data processing class because ContentDescription class was gettgin too large.
