+++ src/AudioAnalysisTools/ContentDescriptionTools/ContentTypes/WindLight1.cs
@@ -38,12 +38,12 @@ namespace AudioAnalysisTools.ContentDescriptionTools.ContentTypes
 
         public static KeyValuePair<string, double> GetContent(Dictionary<string, double[]> oneMinuteOfIndices)
         {
-            var reducedIndices = ContentDescription.ReduceIndicesByFactor(oneMinuteOfIndices, ReductionFactor);
+            var reducedIndices = DataProcessing.ReduceIndicesByFactor(oneMinuteOfIndices, ReductionFactor);
             reducedIndices.Remove("ENT");
             LightWindTemplate.Remove("ENT");
 
-            var oneMinuteVector = ContentDescription.ConvertDictionaryToVector(reducedIndices);
-            var templateVector = ContentDescription.ConvertDictionaryToVector(LightWindTemplate);
+            var oneMinuteVector = DataProcessing.ConvertDictionaryToVector(reducedIndices);
+            var templateVector = DataProcessing.ConvertDictionaryToVector(LightWindTemplate);
 
             var distance = DataTools.EuclideanDistance(templateVector, oneMinuteVector);
 
@@ -54,9 +54,9 @@ namespace AudioAnalysisTools.ContentDescriptionTools.ContentTypes
 
         public static Dictionary<string, double[]> GetTemplate(DirectoryInfo dir)
         {
-            var dictionaryOfIndices = ContentDescription.ReadIndexMatrices(dir, BaseName);
-            var windIndices = ContentDescription.AverageIndicesOverMinutes(dictionaryOfIndices, StartRowId, EndRowId);
-            var reducedIndices = ContentDescription.ReduceIndicesByFactor(windIndices, ReductionFactor);
+            var dictionaryOfIndices = DataProcessing.ReadIndexMatrices(dir, BaseName);
+            var windIndices = DataProcessing.AverageIndicesOverMinutes(dictionaryOfIndices, StartRowId, EndRowId);
+            var reducedIndices = DataProcessing.ReduceIndicesByFactor(windIndices, ReductionFactor);
             return reducedIndices;
         }
